Added a fullscreen button
Also changed `update` to `upsert` in the server/methods.js file, so
that users can create a duplicate using the published fields by
changing the `_id`.  Devs will have to be careful with this, I guess,
or they might end up with their data in a weird state (important
unpublished fields could be missing, breaking app functionality).

Prior to this change, an update with a changed _id would just flick
back to old version and any other edits would be lost too.

You may want to change it back to `update` rather than `upsert` — have
a think about it.